+++ src/services/context/context.spec.ts
@@ -74,50 +74,13 @@ describe('Context Service', () => {
         expect(theSpy.calls.argsFor(0)[0]).toBe(`/organizations/${orgName}/${appBasePath}`);
     });
 
-    it('Gets the org from the user orgs', () => {
-        const orgName = 'orgFromRoles';
-        const user = 'user@apigee.com';
-        const btoaString = `{"email":"${user}"}`;
-        const btoaVal = btoa(btoaString);
-        window.cookieString = `access_token=.${btoaVal}.`;
-        spyOn(loc, 'go');
-        spyOn(client, 'get').and.callThrough();
-        let theSpy = client.get as any;
-        let theLocSpy = loc.go as any;
-        client.on(`/users/${user}/userroles`, { role: [{ organization: orgName}]});
+    it('Redirects to no-org page if organization was not in url nor in local storage', () => {
+        spyOn(loc, 'go').and.callThrough();
+        const theSpy = loc.go as any;
         const context = service.getContext();
         expect(context.orgName).toBeUndefined();
+        expect(theSpy).toHaveBeenCalled();
         expect(theSpy.calls.count()).toBe(1);
-        expect(theSpy.calls.argsFor(0)[0]).toBe(`/users/${user}/userroles`);
-        expect(loc.go).toHaveBeenCalled();
-        expect(theLocSpy.calls.count()).toBe(1);
-        expect(theLocSpy.calls.argsFor(0)[0]).toBe(`/organizations/orgFromRoles/${appBasePath}`);
+        expect(theSpy.calls.argsFor(0)[0]).toBe(`/no-org`);
     });
-
-    // org page does not exist yet, but we should redirect to it
-    it('Should not redirect if the user does not have any roles at any org', () => {
-        const user = 'user@apigee.com';
-        const btoaString = `{"email":"${user}"}`;
-        const btoaVal = btoa(btoaString);
-        window.cookieString = `access_token=.${btoaVal}.`;
-        spyOn(loc, 'go');
-        spyOn(client, 'get').and.callThrough();
-        let theSpy = client.get as any;
-        client.on(`/users/${user}/userroles`, {}); // empty roles
-        const context = service.getContext();
-        expect(context.orgName).toBeUndefined();
-        expect(loc.go).not.toHaveBeenCalled();
-        expect(theSpy.calls.count()).toBe(1);
-        expect(theSpy.calls.argsFor(0)[0]).toBe(`/users/${user}/userroles`);
-    });
-
-    it('Uses http to force redirection', () => {
-        spyOn(client, 'get').and.callThrough();
-        let theSpy = client.get as any;
-        const context = service.getContext();
-        expect(context.orgName).toBeUndefined();
-        expect(theSpy.calls.count()).toBe(1);
-        expect(theSpy.calls.argsFor(0)[0]).toBe('/userinfo');
-    });
-
 });